Helpful API Reference
I think the NYT Most Popular API is actually going to be best for our
purposes. The URIs here show you how to request the most viewed
articles in any NYT section within the last 1, 7, or 30 days. I put
this together because we’ll all need to become acquainted with the API
to test our code.